requirements.txt
azure-functions
psycopg2-binary==2.9.10